{"schema":"decodex/commit/1","summary":"Split operational evidence reporting modules","authority":"manual"}
