+++ apps/elf-eval/src/bin/real_world_job_benchmark/operational/tags.rs
@@ -1,0 +1,5 @@
+use crate::RealWorldJob;
+
+pub(in crate::operational) fn job_has_tag(job: &RealWorldJob, tag: &str) -> bool {
+	job.tags.iter().any(|candidate| candidate == tag)
+}
